+++ tt_esp32controller/src/state.h
@@ -7,9 +7,9 @@
 
 // Firmware-Version (entspricht dem CHANGELOG)
 #ifdef SIM
-#define FW  "Firmware V4.8.0 (SIM)"
+#define FW  "Firmware V4.8.1 (SIM)"
 #else
-#define FW  "Firmware V4.8.0"
+#define FW  "Firmware V4.8.1"
 #endif
 
 // System